Run Enrichment Analysis (Over-representation Analysis)

Run enrichment analysis on a list of differentially expressed genes (DEGs)

Usage

RunEnrichment(
  degs,
  deg = "p_val_adj < 0.05",
  dbs = "KEGG_2021_Human",
  cache = NULL,
  error = TRUE,
  site = "Enrichr",
  ...
)

Arguments

degs

A data frame of DEGs, typically generated by RunSeuratDEAnalysis

deg

A string specifying the condition for a gene to be considered as differentially expressed

dbs

A string specifying the database to use for enrichment analysis A list of available databases can be found at https://maayanlab.cloud/Enrichr/#libraries

cache

A string specifying the directory to cache the results Set to FALSE to disable caching

error

A boolean specifying whether to throw an error if the enrichment analysis fails

site

A string specifying the Enrichr site to use

...

Additional arguments to pass to RunEnrichment

Examples

degs <- suppressWarnings(RunSeuratDEAnalysis(SeuratObject::pbmc_small, "groups", "g1", "g2"))
RunEnrichment(degs, deg = "abs(avg_log2FC) > 1")
#> Connection changed to https://maayanlab.cloud/Enrichr/
#> Connection is Live!
#> Uploading data to Enrichr... Uploading data to Enrichr... 
#> Error in f(init, x[[i]]): is.request(y) is not TRUE