RunSeuratDEAnalysis

Run differential expression analysis on a Seurat object

Usage

RunSeuratDEAnalysis(
  object,
  group_by,
  ident_1 = NULL,
  ident_2 = NULL,
  assay = NULL,
  subset = NULL,
  cache = NULL,
  error = TRUE,
  ...
)

Arguments

object

Seurat object

group_by

Column name in meta.data to group cells by

ident_1

Identity of the first group of cells

ident_2

Identity of the second group of cells

assay

Assay to use for analysis

subset

Subset of cells to use for analysis It should be a string of expression to pass to dplyr::filter function

cache

Directory to cache the results. Set to FALSE to disable caching The results of the analysis will be saved in this directory Anything changed in the object or arguments will trigger a re-run

error

Whether to raise an error if the analysis fails Otherwise, return an empty data frame

...

Additional arguments to pass to Seurat::FindMarkers()

Examples

RunSeuratDEAnalysis(SeuratObject::pbmc_small, "groups", "g1", "g2")
#> Warning: `PackageCheck()` was deprecated in SeuratObject 5.0.0.
#> ℹ Please use `rlang::check_installed()` instead.
#> ℹ The deprecated feature was likely used in the Seurat package.
#>   Please report the issue at <https://github.com/satijalab/seurat/issues>.
#>                    p_val avg_log2FC pct.1 pct.2 p_val_adj          gene groups
#> NOSIP         0.02870319  1.2224797 0.455 0.194         1         NOSIP     NA
#> LYAR          0.02920961  0.8205500 0.159 0.389         1          LYAR     NA
#> SNHG7         0.03189563  0.9583741 0.295 0.083         1         SNHG7     NA
#> TMUB1         0.03592718  1.7965894 0.182 0.028         1         TMUB1     NA
#> ZNF330        0.03911949  9.7197369 0.114 0.000         1        ZNF330     NA
#> FUOM          0.03911949  7.6369685 0.114 0.000         1          FUOM     NA
#> CCL4          0.04262696  3.6666376 0.205 0.056         1          CCL4     NA
#> LINC00936     0.04918625  0.7604677 0.386 0.167         1     LINC00936     NA
#> TSPO          0.05000053  0.8252648 0.659 0.361         1          TSPO     NA
#> IFI6          0.05588343  0.8570195 0.432 0.222         1          IFI6     NA
#> VDAC3         0.06665008  3.6742886 0.318 0.139         1         VDAC3     NA
#> SPON2         0.07526141  1.7237839 0.273 0.111         1         SPON2     NA
#> LGALS2        0.08238383  1.4537375 0.341 0.167         1        LGALS2     NA
#> RPL7L1        0.08259804 -2.5594238 0.091 0.222         1        RPL7L1     NA
#> TAF7          0.08349570  1.5772505 0.341 0.167         1          TAF7     NA
#> FCGR3A        0.08436567  1.1275295 0.432 0.250         1        FCGR3A     NA
#> LGALS9        0.08733037  1.1241969 0.273 0.111         1        LGALS9     NA
#> GZMB          0.08733037  1.1698448 0.273 0.111         1          GZMB     NA
#> CLEC10A       0.09071524  1.6996524 0.182 0.056         1       CLEC10A     NA
#> FCER1A        0.09397257  1.9329064 0.182 0.056         1        FCER1A     NA
#> LRRC25        0.09529164  0.4200849 0.273 0.111         1        LRRC25     NA
#> BANK1         0.09545145  2.0219200 0.136 0.028         1         BANK1     NA
#> GSTP1         0.09628003  0.7713358 0.591 0.444         1         GSTP1     NA
#> SIT1          0.09934503 -3.6034439 0.023 0.111         1          SIT1     NA
#> ARHGDIA       0.10084609 -0.4601009 0.432 0.250         1       ARHGDIA     NA
#> HLA-DMA       0.10813435  1.0195709 0.386 0.222         1       HLA-DMA     NA
#> GRN           0.11667166  0.5547349 0.409 0.222         1           GRN     NA
#> RUFY1         0.11768586 -0.9810894 0.091 0.222         1         RUFY1     NA
#> MS4A7         0.12277557  0.5415504 0.227 0.083         1         MS4A7     NA
#> IFI30         0.13211385  0.4138543 0.341 0.167         1         IFI30     NA
#> RGS1          0.13634254  2.5506416 0.159 0.056         1          RGS1     NA
#> THYN1         0.14336601  1.5281674 0.205 0.083         1         THYN1     NA
#> PARVB         0.15081270 -1.4496990 0.068 0.167         1         PARVB     NA
#> PCMT1         0.15207496  3.3593272 0.295 0.167         1         PCMT1     NA
#> FCGRT         0.16468927  0.9702779 0.409 0.250         1         FCGRT     NA
#> BID           0.17214063  1.1827776 0.386 0.222         1           BID     NA
#> GNLY          0.17975936  1.3725474 0.295 0.167         1          GNLY     NA
#> FCER1G        0.18735817  0.6490939 0.636 0.472         1        FCER1G     NA
#> TAGAP         0.18771440 -0.4875612 0.114 0.222         1         TAGAP     NA
#> BLVRA         0.19048552 -1.1214685 0.227 0.333         1         BLVRA     NA
#> COTL1         0.19393336  0.4875465 0.750 0.583         1         COTL1     NA
#> HLA-DQA1      0.21368858  0.9605902 0.386 0.250         1      HLA-DQA1     NA
#> VPS28         0.21488446 -1.2334213 0.318 0.444         1         VPS28     NA
#> GIMAP1        0.22134291  0.9678929 0.409 0.250         1        GIMAP1     NA
#> PRF1          0.22555711  0.3097331 0.295 0.167         1          PRF1     NA
#> SAFB2         0.23520876  2.4093824 0.091 0.028         1         SAFB2     NA
#> EIF3G         0.23894997  0.7393618 0.591 0.417         1         EIF3G     NA
#> NRBP1         0.24206600 -2.2937805 0.091 0.167         1         NRBP1     NA
#> HLA-DQA2      0.24489452  0.6449536 0.250 0.139         1      HLA-DQA2     NA
#> SCO2          0.24808905  0.8095112 0.182 0.083         1          SCO2     NA
#> PTPN22        0.25724622 -2.1902685 0.045 0.111         1        PTPN22     NA
#> PPP3CC        0.26106405  0.7908963 0.136 0.056         1        PPP3CC     NA
#> ZNF76         0.26360996  1.2855973 0.091 0.028         1         ZNF76     NA
#> CD1C          0.26360996  1.5571143 0.091 0.028         1          CD1C     NA
#> WARS          0.26808657 -1.3804565 0.159 0.250         1          WARS     NA
#> EPC1          0.27173354  0.8665448 0.205 0.111         1          EPC1     NA
#> TCL1A         0.27307223 -1.8951067 0.068 0.139         1         TCL1A     NA
#> CCDC104       0.27358239  0.9099128 0.091 0.028         1       CCDC104     NA
#> GZMH          0.28136356  1.6970640 0.159 0.083         1          GZMH     NA
#> HLA-DRA       0.28215715  0.2258141 0.727 0.639         1       HLA-DRA     NA
#> RBP7          0.28966268 -1.1441332 0.068 0.139         1          RBP7     NA
#> RGS2          0.29358016 -0.6789153 0.409 0.278         1          RGS2     NA
#> MPHOSPH6      0.29667195  1.4082929 0.159 0.083         1      MPHOSPH6     NA
#> TREML1        0.29820733 -1.3982890 0.068 0.139         1        TREML1     NA
#> TCF7          0.30419782 -0.9406441 0.114 0.194         1          TCF7     NA
#> PTCRA         0.30440015 -0.2643378 0.045 0.111         1         PTCRA     NA
#> BLOC1S4       0.30851235 -1.2148305 0.091 0.167         1       BLOC1S4     NA
#> CD9           0.30851235  0.1021292 0.091 0.167         1           CD9     NA
#> TNFRSF1B      0.30987432  0.8275763 0.227 0.139         1      TNFRSF1B     NA
#> TPM4          0.31495381  0.4273507 0.409 0.278         1          TPM4     NA
#> LST1          0.32907845  0.1394113 0.500 0.361         1          LST1     NA
#> FAM96A        0.33150327 -0.4500508 0.205 0.111         1        FAM96A     NA
#> SERPINA1      0.33384685  0.8580429 0.341 0.250         1      SERPINA1     NA
#> C12orf75      0.34229821 -0.5783239 0.114 0.194         1      C12orf75     NA
#> HLA-DRB1      0.34667535  0.3435079 0.568 0.444         1      HLA-DRB1     NA
#> HLA-DPA1      0.35432386  0.6422033 0.545 0.500         1      HLA-DPA1     NA
#> CARD16        0.35475812  0.1396680 0.409 0.278         1        CARD16     NA
#> HLA-DRB5      0.35956376  0.1505533 0.523 0.361         1      HLA-DRB5     NA
#> ZFP36L1       0.36236507 -2.0075334 0.136 0.194         1       ZFP36L1     NA
#> RHOC          0.36735744 -0.2360902 0.409 0.250         1          RHOC     NA
#> CD247         0.37584672  0.4950876 0.250 0.167         1         CD247     NA
#> CFP           0.38055093 -0.4032603 0.364 0.250         1           CFP     NA
#> GZMM          0.38163435  0.7040406 0.273 0.194         1          GZMM     NA
#> SAT1          0.38637069  0.1496136 0.773 0.639         1          SAT1     NA
#> TALDO1        0.38799801 -1.4821323 0.500 0.500         1        TALDO1     NA
#> YWHAB         0.40132815  0.3123683 0.523 0.528         1         YWHAB     NA
#> FPR1          0.40338593 -1.3374693 0.136 0.194         1          FPR1     NA
#> TNFAIP8       0.41017526 -0.6427122 0.250 0.306         1       TNFAIP8     NA
#> TTC38         0.41239096  1.2736847 0.136 0.083         1         TTC38     NA
#> HVCN1         0.41463077 -2.2939901 0.114 0.167         1         HVCN1     NA
#> CCL5          0.41474716 -0.9043819 0.432 0.472         1          CCL5     NA
#> LYZ           0.42215674  0.2979607 0.682 0.583         1           LYZ     NA
#> VSTM1         0.42420482 -0.2914488 0.114 0.056         1         VSTM1     NA
#> CD8A          0.42421612 -0.4965425 0.114 0.056         1          CD8A     NA
#> LAMP1         0.42538075 -0.5861493 0.205 0.278         1         LAMP1     NA
#> CD200         0.43621807 -2.5445648 0.023 0.056         1         CD200     NA
#> CD19          0.44237088 -1.0048143 0.068 0.028         1          CD19     NA
#> DLGAP1-AS1    0.44237088  0.5505892 0.068 0.028         1    DLGAP1-AS1     NA
#> FCER2         0.45379221 -1.7096009 0.045 0.083         1         FCER2     NA
#> FCRLA         0.45379221 -2.2473152 0.045 0.083         1         FCRLA     NA
#> TMEM40        0.45934038 -1.4708196 0.068 0.111         1        TMEM40     NA
#> ASGR1         0.46444068  1.3781397 0.136 0.083         1         ASGR1     NA
#> MMADHC        0.46691934 -1.2601668 0.205 0.250         1        MMADHC     NA
#> RP11-693J15.5 0.46780299 -1.9053842 0.045 0.083         1 RP11-693J15.5     NA
#> IGLL5         0.47162751 -0.2402105 0.023 0.056         1         IGLL5     NA
#> ANXA2         0.47261188  0.1269778 0.614 0.444         1         ANXA2     NA
#> S1PR4         0.47993990  1.7654906 0.273 0.194         1         S1PR4     NA
#> MYL9          0.48363363 -0.9581845 0.068 0.111         1          MYL9     NA
#> COPS6         0.49620744 -0.8663350 0.273 0.194         1         COPS6     NA
#> FCN1          0.49830814 -0.5499683 0.386 0.417         1          FCN1     NA
#> CD68          0.49979451  0.4068668 0.295 0.222         1          CD68     NA
#> SP100         0.50025570 -0.5112157 0.295 0.194         1         SP100     NA
#> CD79A         0.50287239 -0.6183097 0.136 0.194         1         CD79A     NA
#> GPX1          0.51192812 -0.6122229 0.682 0.639         1          GPX1     NA
#> CCR7          0.51836193  1.9506907 0.091 0.056         1          CCR7     NA
#> ADAR          0.52877864 -0.1155165 0.182 0.222         1          ADAR     NA
#> ACRBP         0.53424169  0.8715893 0.068 0.111         1         ACRBP     NA
#> FGFBP2        0.53811724  1.3991481 0.250 0.222         1        FGFBP2     NA
#> TYROBP        0.54258037  0.1047340 0.636 0.500         1        TYROBP     NA
#> GP9           0.54308166 -0.2629969 0.091 0.139         1           GP9     NA
#> ODC1          0.54704146 -0.8462178 0.182 0.222         1          ODC1     NA
#> MS4A1         0.54893336 -1.7447339 0.136 0.167         1         MS4A1     NA
#> KLRD1         0.55544738  0.5675050 0.159 0.111         1         KLRD1     NA
#> LY86          0.56056635  0.1599974 0.250 0.194         1          LY86     NA
#> NFKBIA        0.56488054  0.6054230 0.591 0.472         1        NFKBIA     NA
#> NKG7          0.56823957  0.2819263 0.477 0.389         1          NKG7     NA
#> DNAJB1        0.57140408 -0.2655006 0.227 0.278         1        DNAJB1     NA
#> LCK           0.57352870  0.3533948 0.318 0.250         1           LCK     NA
#> MPEG1         0.57892640 -0.7608735 0.205 0.139         1         MPEG1     NA
#> SMCO4         0.58895599 -0.3460177 0.205 0.139         1         SMCO4     NA
#> IL1B          0.60418281  1.0822189 0.205 0.167         1          IL1B     NA
#> AIF1          0.60882656  0.1259509 0.545 0.444         1          AIF1     NA
#> HNRNPA3       0.60921704 -0.1623689 0.409 0.306         1       HNRNPA3     NA
#> AKR1C3        0.61597272  0.9619857 0.114 0.083         1        AKR1C3     NA
#> CD14          0.61951529 -1.2948842 0.159 0.194         1          CD14     NA
#> CD3E          0.62522103  0.4977330 0.341 0.278         1          CD3E     NA
#> GNG11         0.63742523 -1.1347009 0.114 0.139         1         GNG11     NA
#> CLIC3         0.64234460  0.9081385 0.114 0.083         1         CLIC3     NA
#> CLU           0.64949087 -0.9996102 0.114 0.139         1           CLU     NA
#> CTSW          0.65028217 -0.3922628 0.341 0.250         1          CTSW     NA
#> STX10         0.65074656 -0.7783398 0.159 0.194         1         STX10     NA
#> GZMK          0.66243894 -1.7574529 0.091 0.111         1          GZMK     NA
#> IL7R          0.66674361  0.2082191 0.341 0.333         1          IL7R     NA
#> PTGDR         0.66918992 -2.7669811 0.114 0.083         1         PTGDR     NA
#> S100B         0.67009042  4.1497946 0.045 0.028         1         S100B     NA
#> LILRA3        0.67596773 -1.9791945 0.091 0.111         1        LILRA3     NA
#> DNAJC2        0.67885156 -0.1544783 0.182 0.139         1        DNAJC2     NA
#> GYPC          0.68005668 -0.8651684 0.250 0.278         1          GYPC     NA
#> MYO1G         0.68065858  0.6695113 0.409 0.333         1         MYO1G     NA
#> IL2RB         0.68277191  0.2945914 0.114 0.083         1         IL2RB     NA
#> PPP1R18       0.68939101  0.3244785 0.318 0.278         1       PPP1R18     NA
#> NT5C          0.70343211 -1.2694963 0.227 0.167         1          NT5C     NA
#> S100A11       0.70479142  0.1499056 0.591 0.528         1       S100A11     NA
#> XBP1          0.71198923  0.1016697 0.318 0.250         1          XBP1     NA
#> TMEM204       0.71342360  0.4482229 0.045 0.028         1       TMEM204     NA
#> S100A8        0.71501151 -1.2624184 0.409 0.306         1        S100A8     NA
#> NCOA4         0.73091444 -0.8719660 0.250 0.194         1         NCOA4     NA
#> PGRMC1        0.73809476 -0.4498943 0.114 0.083         1        PGRMC1     NA
#> CTSB          0.73923287 -0.6614962 0.318 0.250         1          CTSB     NA
#> IGFBP7        0.75069884 -0.5043795 0.205 0.222         1        IGFBP7     NA
#> CD79B         0.76680522 -1.2150744 0.295 0.278         1         CD79B     NA
#> SATB1         0.76687418 -3.0600459 0.068 0.083         1         SATB1     NA
#> RP11-290F20.3 0.76832326  0.1220792 0.136 0.111         1 RP11-290F20.3     NA
#> HLA-DMB       0.77306125 -0.3343191 0.341 0.278         1       HLA-DMB     NA
#> CD2           0.77334050  0.5941658 0.159 0.139         1           CD2     NA
#> PIK3IP1       0.77930880  1.4533085 0.159 0.139         1       PIK3IP1     NA
#> GZMA          0.78049937 -0.1186839 0.227 0.250         1          GZMA     NA
#> PMPCB         0.78121854  0.1265535 0.136 0.111         1         PMPCB     NA
#> KIAA0125      0.80780244 -1.8552812 0.045 0.056         1      KIAA0125     NA
#> CYB561A3      0.80780244 -1.8289438 0.045 0.056         1      CYB561A3     NA
#> CD180         0.80780244 -1.5002871 0.045 0.056         1         CD180     NA
#> HNRNPH1       0.82024117 -0.2902894 0.136 0.111         1       HNRNPH1     NA
#> IL32          0.82066860 -0.5310219 0.295 0.306         1          IL32     NA
#> ASNSD1        0.82403692 -0.2054350 0.068 0.083         1        ASNSD1     NA
#> MAL           0.83229313  0.2357011 0.068 0.083         1           MAL     NA
#> LINC00926     0.84467238 -0.3201780 0.068 0.056         1     LINC00926     NA
#> PPBP          0.85072316 -0.8161701 0.159 0.167         1          PPBP     NA
#> HNRNPF        0.85270253 -0.1318392 0.432 0.361         1        HNRNPF     NA
#> IL17RA        0.86275280 -1.1075541 0.068 0.056         1        IL17RA     NA
#> SRSF7         0.87608406 -1.5599110 0.500 0.472         1         SRSF7     NA
#> POP7          0.87628706  1.5192430 0.159 0.139         1          POP7     NA
#> PPAPDC1B      0.88090539 -0.8740292 0.068 0.056         1      PPAPDC1B     NA
#> NRGN          0.88965687  0.1178269 0.182 0.167         1          NRGN     NA
#> XCL2          0.89794707  0.3356016 0.091 0.083         1          XCL2     NA
#> CFD           0.90550039 -0.4594260 0.273 0.250         1           CFD     NA
#> ITGA2B        0.90877741 -1.4419882 0.114 0.111         1        ITGA2B     NA
#> LDHB          0.90895712 -0.1769172 0.523 0.444         1          LDHB     NA
#> SDPR          0.91637557 -0.7586314 0.136 0.139         1          SDPR     NA
#> EIF4A2        0.92216620 -0.1168652 0.523 0.500         1        EIF4A2     NA
#> SPARC         0.92920397 -0.6536246 0.136 0.139         1         SPARC     NA
#> S100A9        0.93413505 -0.7924142 0.386 0.389         1        S100A9     NA
#> EAF2          0.93677400 -2.6687209 0.114 0.111         1          EAF2     NA
#> KHDRBS1       0.94760219 -2.8736243 0.182 0.167         1       KHDRBS1     NA
#> C5AR1         0.95034695 -0.9861914 0.159 0.139         1         C5AR1     NA
#> HLA-DPB1      0.95531585  0.2193854 0.523 0.556         1      HLA-DPB1     NA
#> TYMP          0.95715084 -0.1454613 0.432 0.417         1          TYMP     NA
#> CRBN          0.95942509 -0.8320513 0.273 0.222         1          CRBN     NA
#> CTSS          0.96412214  0.1540970 0.591 0.611         1          CTSS     NA
#> KLRG1         0.96484898 -0.3207581 0.114 0.111         1         KLRG1     NA
#> MS4A6A        0.98895497 -0.5252983 0.205 0.194         1        MS4A6A     NA
#> NGFRAP1       0.99296761 -0.6881485 0.114 0.111         1       NGFRAP1     NA
#> CA2           0.99296761  0.8072664 0.114 0.111         1           CA2     NA
#> CD3D          0.99517096 -0.5121695 0.295 0.278         1          CD3D     NA
#> RARRES3       0.99556237  0.1837819 0.364 0.389         1       RARRES3     NA
#>               diff_pct
#> NOSIP            0.261
#> LYAR            -0.230
#> SNHG7            0.212
#> TMUB1            0.154
#> ZNF330           0.114
#> FUOM             0.114
#> CCL4             0.149
#> LINC00936        0.219
#> TSPO             0.298
#> IFI6             0.210
#> VDAC3            0.179
#> SPON2            0.162
#> LGALS2           0.174
#> RPL7L1          -0.131
#> TAF7             0.174
#> FCGR3A           0.182
#> LGALS9           0.162
#> GZMB             0.162
#> CLEC10A          0.126
#> FCER1A           0.126
#> LRRC25           0.162
#> BANK1            0.108
#> GSTP1            0.147
#> SIT1            -0.088
#> ARHGDIA          0.182
#> HLA-DMA          0.164
#> GRN              0.187
#> RUFY1           -0.131
#> MS4A7            0.144
#> IFI30            0.174
#> RGS1             0.103
#> THYN1            0.122
#> PARVB           -0.099
#> PCMT1            0.128
#> FCGRT            0.159
#> BID              0.164
#> GNLY             0.128
#> FCER1G           0.164
#> TAGAP           -0.108
#> BLVRA           -0.106
#> COTL1            0.167
#> HLA-DQA1         0.136
#> VPS28           -0.126
#> GIMAP1           0.159
#> PRF1             0.128
#> SAFB2            0.063
#> EIF3G            0.174
#> NRBP1           -0.076
#> HLA-DQA2         0.111
#> SCO2             0.099
#> PTPN22          -0.066
#> PPP3CC           0.080
#> ZNF76            0.063
#> CD1C             0.063
#> WARS            -0.091
#> EPC1             0.094
#> TCL1A           -0.071
#> CCDC104          0.063
#> GZMH             0.076
#> HLA-DRA          0.088
#> RBP7            -0.071
#> RGS2             0.131
#> MPHOSPH6         0.076
#> TREML1          -0.071
#> TCF7            -0.080
#> PTCRA           -0.066
#> BLOC1S4         -0.076
#> CD9             -0.076
#> TNFRSF1B         0.088
#> TPM4             0.131
#> LST1             0.139
#> FAM96A           0.094
#> SERPINA1         0.091
#> C12orf75        -0.080
#> HLA-DRB1         0.124
#> HLA-DPA1         0.045
#> CARD16           0.131
#> HLA-DRB5         0.162
#> ZFP36L1         -0.058
#> RHOC             0.159
#> CD247            0.083
#> CFP              0.114
#> GZMM             0.079
#> SAT1             0.134
#> TALDO1           0.000
#> YWHAB           -0.005
#> FPR1            -0.058
#> TNFAIP8         -0.056
#> TTC38            0.053
#> HVCN1           -0.053
#> CCL5            -0.040
#> LYZ              0.099
#> VSTM1            0.058
#> CD8A             0.058
#> LAMP1           -0.073
#> CD200           -0.033
#> CD19             0.040
#> DLGAP1-AS1       0.040
#> FCER2           -0.038
#> FCRLA           -0.038
#> TMEM40          -0.043
#> ASGR1            0.053
#> MMADHC          -0.045
#> RP11-693J15.5   -0.038
#> IGLL5           -0.033
#> ANXA2            0.170
#> S1PR4            0.079
#> MYL9            -0.043
#> COPS6            0.079
#> FCN1            -0.031
#> CD68             0.073
#> SP100            0.101
#> CD79A           -0.058
#> GPX1             0.043
#> CCR7             0.035
#> ADAR            -0.040
#> ACRBP           -0.043
#> FGFBP2           0.028
#> TYROBP           0.136
#> GP9             -0.048
#> ODC1            -0.040
#> MS4A1           -0.031
#> KLRD1            0.048
#> LY86             0.056
#> NFKBIA           0.119
#> NKG7             0.088
#> DNAJB1          -0.051
#> LCK              0.068
#> MPEG1            0.066
#> SMCO4            0.066
#> IL1B             0.038
#> AIF1             0.101
#> HNRNPA3          0.103
#> AKR1C3           0.031
#> CD14            -0.035
#> CD3E             0.063
#> GNG11           -0.025
#> CLIC3            0.031
#> CLU             -0.025
#> CTSW             0.091
#> STX10           -0.035
#> GZMK            -0.020
#> IL7R             0.008
#> PTGDR            0.031
#> S100B            0.017
#> LILRA3          -0.020
#> DNAJC2           0.043
#> GYPC            -0.028
#> MYO1G            0.076
#> IL2RB            0.031
#> PPP1R18          0.040
#> NT5C             0.060
#> S100A11          0.063
#> XBP1             0.068
#> TMEM204          0.017
#> S100A8           0.103
#> NCOA4            0.056
#> PGRMC1           0.031
#> CTSB             0.068
#> IGFBP7          -0.017
#> CD79B            0.017
#> SATB1           -0.015
#> RP11-290F20.3    0.025
#> HLA-DMB          0.063
#> CD2              0.020
#> PIK3IP1          0.020
#> GZMA            -0.023
#> PMPCB            0.025
#> KIAA0125        -0.011
#> CYB561A3        -0.011
#> CD180           -0.011
#> HNRNPH1          0.025
#> IL32            -0.011
#> ASNSD1          -0.015
#> MAL             -0.015
#> LINC00926        0.012
#> PPBP            -0.008
#> HNRNPF           0.071
#> IL17RA           0.012
#> SRSF7            0.028
#> POP7             0.020
#> PPAPDC1B         0.012
#> NRGN             0.015
#> XCL2             0.008
#> CFD              0.023
#> ITGA2B           0.003
#> LDHB             0.079
#> SDPR            -0.003
#> EIF4A2           0.023
#> SPARC           -0.003
#> S100A9          -0.003
#> EAF2             0.003
#> KHDRBS1          0.015
#> C5AR1            0.020
#> HLA-DPB1        -0.033
#> TYMP             0.015
#> CRBN             0.051
#> CTSS            -0.020
#> KLRG1            0.003
#> MS4A6A           0.011
#> NGFRAP1          0.003
#> CA2              0.003
#> CD3D             0.017
#> RARRES3         -0.025