Internal helper function to compare the size of two groups.

Internal helper function to compare the size of two groups.

Usage

.size_compare(
  df,
  group_by,
  idents,
  id,
  fun,
  compare = ".n",
  split_by = NULL,
  order = "desc(sum)",
  subset = NULL,
  return_type = c("uids", "ids", "subdf", "df", "interdf"),
  include_zeros = FALSE
)

Arguments

df

The data frame. Use . if the function is called in a dplyr pipe.

group_by

The column name in the data frame to group the entities. It could be a quoted string or a bare variable, and defines the groups of entities for comparison.

idents

The values in compare to compare. It could be either an an integer or a vector. If it is an integer, the number of values in compare must be the same as the integer for the id to be regarded as paired. If it is a vector, the values in compare must be the same as the values in idents for the id to be regarded as paired.

id

The column name in df for the groups.

fun

The way to compare between groups. Either "more_than", "less_than", "emerged" or "vanished".

compare

The column name in df to compare the values for each group. It could be either a numeric column or .n to compare the number of entities in each group. If a column is passed, the values in the column must be numeric and the same in each group. This won't be checked.

split_by

A column name (without quotes) in metadata to split the cells.

order

An expression to order the intermediate data frame before returning the final result. Default is NULL. It does not work for subdf and df.

subset

An expression to subset the entities, will be passed to dplyr::filter(). Default is TRUE (no filtering).

return_type

The type of the returned value. Default is uids. It could be one of

uids: return the unique ids of the selected entities

ids: return the ids of all entities in the same order as in df, where the non-selected ids will be NA

subdf: return a subset of df with the selected entities

df: return the original df with a new logical column .out to mark the selected entities

interdf: return the intermediate data frame with the id column, <compare>, predicate and the split_by column if provided.

include_zeros

Whether to include the zero entities in the other group for more_than and less_than comparisons. Default is FALSE. By default, the zero entities will be excluded, meaning that the entities must exist in both groups to be selected.

Value

A vector of ids or a data frame with the selected entities.