Run Enrichment Analysis (Over-representation Analysis)

Run enrichment analysis on a list of differentially expressed genes (DEGs)

Usage

RunEnrichment(
  degs,
  deg = "p_val_adj < 0.05",
  dbs = "KEGG_2021_Human",
  cache = NULL,
  error = TRUE,
  site = "Enrichr",
  ...
)

Arguments

degs

A data frame of DEGs, typically generated by RunSeuratDEAnalysis

deg

A string specifying the condition for a gene to be considered as differentially expressed

dbs

A string specifying the database to use for enrichment analysis A list of available databases can be found at https://maayanlab.cloud/Enrichr/#libraries

cache

A string specifying the directory to cache the results Set to FALSE to disable caching

error

A boolean specifying whether to throw an error if the enrichment analysis fails

site

A string specifying the Enrichr site to use

...

Additional arguments to pass to RunEnrichment

Examples

# \donttest{
degs <- suppressWarnings(RunSeuratDEAnalysis(SeuratObject::pbmc_small,
 "groups", "g1", "g2"))
RunEnrichment(degs, deg = "abs(avg_log2FC) > 1")
#> Connection changed to https://maayanlab.cloud/Enrichr/
#> Connection is Live!
#> Uploading data to Enrichr... Done.
#>   Querying KEGG_2021_Human... Done.
#> Parsing results... Done.
#>                                                              Term Overlap
#> 1                                      Hematopoietic cell lineage    8/99
#> 2                                       Graft-versus-host disease    3/42
#> 3                                        Type I diabetes mellitus    3/43
#> 4                                                    Tuberculosis   5/180
#> 5                                                   Leishmaniasis    3/77
#> 6                               B cell receptor signaling pathway    3/81
#> 7                                         IL-17 signaling pathway    3/94
#> 8                                 Staphylococcus aureus infection    3/95
#> 9                         Neutrophil extracellular trap formation   4/189
#> 10                                                         Asthma    2/31
#> 11                                                     Amoebiasis   3/102
#> 12                           Toll-like receptor signaling pathway   3/104
#> 13                                   NF-kappa B signaling pathway   3/104
#> 14                                   Epstein-Barr virus infection   4/202
#> 15                                            Allograft rejection    2/38
#> 16                                 Sphingolipid signaling pathway   3/119
#> 17                                Human cytomegalovirus infection   4/225
#> 18                                     Osteoclast differentiation   3/127
#> 19                      Natural killer cell mediated cytotoxicity   3/131
#> 20                                     Autoimmune thyroid disease    2/53
#> 21                                                  Legionellosis    2/57
#> 22                                                      Phagosome   3/152
#> 23                                              Viral myocarditis    2/60
#> 24                                                    Necroptosis   3/159
#> 25                                  Cytosolic DNA-sensing pathway    2/63
#> 26                                     Inflammatory bowel disease    2/65
#> 27                         Cytokine-cytokine receptor interaction   4/295
#> 28                                                    Influenza A   3/172
#> 29                                                      Pertussis    2/76
#> 30                                    Chemokine signaling pathway   3/192
#> 31                                     PI3K-Akt signaling pathway   4/354
#> 32                                      Lipid and atherosclerosis   3/215
#> 33                                           Rheumatoid arthritis    2/93
#> 34  Viral protein interaction with cytokine and cytokine receptor   2/100
#> 35                                      Th17 cell differentiation   2/107
#> 36                                   Systemic lupus erythematosus   2/135
#> 37                         Fluid shear stress and atherosclerosis   2/139
#> 38                                                        Measles   2/139
#> 39                                                      Apoptosis   2/142
#> 40                                      Pentose phosphate pathway    1/30
#> 41                              Non-alcoholic fatty liver disease   2/155
#> 42                                            Cellular senescence   2/156
#> 43                               Herpes simplex virus 1 infection   4/498
#> 44                         Nicotinate and nicotinamide metabolism    1/35
#> 45                                                    Hepatitis B   2/162
#> 46                                        African trypanosomiasis    1/37
#> 47                                       Primary immunodeficiency    1/38
#> 48                        Neuroactive ligand-receptor interaction   3/341
#> 49                                                    Ferroptosis    1/41
#> 50                            NOD-like receptor signaling pathway   2/181
#> 51                           Porphyrin and chlorophyll metabolism    1/43
#> 52                                    Basal transcription factors    1/45
#> 53                                              Alzheimer disease   3/369
#> 54                        Transcriptional misregulation in cancer   2/192
#> 55                Kaposi sarcoma-associated herpesvirus infection   2/193
#> 56                   Intestinal immune network for IgA production    1/48
#> 57                                         Cholesterol metabolism    1/50
#> 58                                                        Malaria    1/50
#> 59                                                 Focal adhesion   2/201
#> 60                                         Rap1 signaling pathway   2/210
#> 61                                          Pyrimidine metabolism    1/56
#> 62                       Human immunodeficiency virus 1 infection   2/212
#> 63                        Human T-cell leukemia virus 1 infection   2/219
#> 64                                         Acute myeloid leukemia    1/67
#> 65                                Fc epsilon RI signaling pathway    1/68
#> 66                                                    Shigellosis   2/246
#> 67                                          p53 signaling pathway    1/73
#> 68                                           Salmonella infection   2/249
#> 69                                      Thyroid hormone synthesis    1/75
#> 70                Arrhythmogenic right ventricular cardiomyopathy    1/77
#> 71                            Antigen processing and presentation    1/78
#> 72                                                RNA degradation    1/79
#> 73                                  Pathways of neurodegeneration   3/475
#> 74                                                  Prion disease   2/273
#> 75                                       ECM-receptor interaction    1/88
#> 76                                              GABAergic synapse    1/89
#> 77                                    Hypertrophic cardiomyopathy    1/90
#> 78                                             Morphine addiction    1/91
#> 79                                         Small cell lung cancer    1/92
#> 80                               Th1 and Th2 cell differentiation    1/92
#> 81                                         Dilated cardiomyopathy    1/96
#> 82                                         MAPK signaling pathway   2/294
#> 83                                          Circadian entrainment    1/97
#> 84                               Fc gamma R-mediated phagocytosis    1/97
#> 85               Inflammatory mediator regulation of TRP channels    1/98
#> 86           AGE-RAGE signaling pathway in diabetic complications   1/100
#> 87                                                 Chagas disease   1/102
#> 88                       C-type lectin receptor signaling pathway   1/104
#> 89                                             Pathways in cancer   3/531
#> 90                              Ribosome biogenesis in eukaryotes   1/108
#> 91                                          TNF signaling pathway   1/112
#> 92                                                  Toxoplasmosis   1/112
#> 93                                           Serotonergic synapse   1/113
#> 94                                            Cholinergic synapse   1/113
#> 95                                          Glutamatergic synapse   1/114
#> 96                                            Platelet activation   1/124
#> 97                                              Purine metabolism   1/129
#> 98                                      Relaxin signaling pathway   1/129
#> 99                                         FoxO signaling pathway   1/131
#> 100                                          Dopaminergic synapse   1/132
#> 101                                 Amyotrophic lateral sclerosis   2/364
#> 102                                      Apelin signaling pathway   1/137
#> 103                                            Yersinia infection   1/137
#> 104                                        Spinocerebellar ataxia   1/143
#> 105                                       Cell adhesion molecules   1/148
#> 106                             Phospholipase D signaling pathway   1/148
#> 107                          Retrograde endocannabinoid signaling   1/148
#> 108                                                   Spliceosome   1/150
#> 109                                                   Hepatitis C   1/157
#> 110                                    cGMP-PKG signaling pathway   1/167
#> 111                                                Tight junction   1/169
#> 112                                                    Alcoholism   1/186
#> 113                                                 RNA transport   1/186
#> 114                         Pathogenic Escherichia coli infection   1/197
#> 115                                          Viral carcinogenesis   1/203
#> 116                                       Diabetic cardiomyopathy   1/203
#> 117                              Regulation of actin cytoskeleton   1/218
#> 118                                           Coronavirus disease   1/232
#> 119                                         Ras signaling pathway   1/232
#> 120                                     Calcium signaling pathway   1/240
#> 121                                             Parkinson disease   1/249
#> 122                                                   Endocytosis   1/252
#> 123                                            Huntington disease   1/306
#> 124                                Human papillomavirus infection   1/331
#>          P.value Adjusted.P.value Old.P.value Old.Adjusted.P.value Odds.Ratio
#> 1   3.864191e-09     4.791597e-07           0                    0 25.6405947
#> 2   5.441297e-04     1.908538e-02           0                    0 20.9536354
#> 3   5.833473e-04     1.908538e-02           0                    0 20.4287671
#> 4   6.156575e-04     1.908538e-02           0                    0  7.9472837
#> 5   3.152458e-03     6.515878e-02           0                    0 11.0236949
#> 6   3.637281e-03     6.515878e-02           0                    0 10.4562698
#> 7   5.516829e-03     6.515878e-02           0                    0  8.9566461
#> 8   5.681341e-03     6.515878e-02           0                    0  8.8588446
#> 9   5.837599e-03     6.515878e-02           0                    0  5.9276276
#> 10  6.170140e-03     6.515878e-02           0                    0 18.5414725
#> 11  6.915482e-03     6.515878e-02           0                    0  8.2295558
#> 12  7.295013e-03     6.515878e-02           0                    0  8.0657806
#> 13  7.295013e-03     6.515878e-02           0                    0  8.0657806
#> 14  7.356637e-03     6.515878e-02           0                    0  5.5347924
#> 15  9.169706e-03     7.580290e-02           0                    0 14.9309309
#> 16  1.053421e-02     7.763074e-02           0                    0  7.0174776
#> 17  1.064292e-02     7.763074e-02           0                    0  4.9529915
#> 18  1.255218e-02     8.647057e-02           0                    0  6.5620857
#> 19  1.363853e-02     8.900938e-02           0                    0  6.3557363
#> 20  1.732811e-02     1.074343e-01           0                    0 10.5315315
#> 21  1.987498e-02     1.138674e-01           0                    0  9.7636364
#> 22  2.020228e-02     1.138674e-01           0                    0  5.4541693
#> 23  2.188161e-02     1.173583e-01           0                    0  9.2572227
#> 24  2.271451e-02     1.173583e-01           0                    0  5.2075869
#> 25  2.396862e-02     1.188844e-01           0                    0  8.8006203
#> 26  2.540362e-02     1.193474e-01           0                    0  8.5203775
#> 27  2.598693e-02     1.193474e-01           0                    0  3.7481863
#> 28  2.781207e-02     1.231677e-01           0                    0  4.8038421
#> 29  3.389234e-02     1.449190e-01           0                    0  7.2498174
#> 30  3.674508e-02     1.518797e-01           0                    0  4.2911503
#> 31  4.584599e-02     1.833840e-01           0                    0  3.1069841
#> 32  4.862084e-02     1.834801e-01           0                    0  3.8211424
#> 33  4.882938e-02     1.834801e-01           0                    0  5.8904069
#> 34  5.555659e-02     2.026182e-01           0                    0  5.4677330
#> 35  6.258860e-02     2.217425e-01           0                    0  5.1014157
#> 36  9.337261e-02     3.200272e-01           0                    0  4.0217435
#> 37  9.807285e-02     3.200272e-01           0                    0  3.9035313
#> 38  9.807285e-02     3.200272e-01           0                    0  3.9035313
#> 39  1.016415e-01     3.231679e-01           0                    0  3.8193050
#> 40  1.080095e-01     3.348295e-01           0                    0  9.1471264
#> 41  1.175045e-01     3.480007e-01           0                    0  3.4924925
#> 42  1.187498e-01     3.480007e-01           0                    0  3.4696385
#> 43  1.212348e-01     3.480007e-01           0                    0  2.1851102
#> 44  1.248558e-01     3.480007e-01           0                    0  7.8000000
#> 45  1.262906e-01     3.480007e-01           0                    0  3.3385135
#> 46  1.315060e-01     3.544944e-01           0                    0  7.3659259
#> 47  1.348124e-01     3.556752e-01           0                    0  7.1664865
#> 48  1.401998e-01     3.621827e-01           0                    0  2.3813731
#> 49  1.446572e-01     3.660713e-01           0                    0  6.6280000
#> 50  1.508671e-01     3.675236e-01           0                    0  2.9812774
#> 51  1.511589e-01     3.675236e-01           0                    0  6.3117460
#> 52  1.576118e-01     3.702298e-01           0                    0  6.0242424
#> 53  1.653864e-01     3.702298e-01           0                    0  2.1960476
#> 54  1.655039e-01     3.702298e-01           0                    0  2.8071124
#> 55  1.668471e-01     3.702298e-01           0                    0  2.7922739
#> 56  1.672006e-01     3.702298e-01           0                    0  5.6388652
#> 57  1.735331e-01     3.710019e-01           0                    0  5.4081633
#> 58  1.735331e-01     3.710019e-01           0                    0  5.4081633
#> 59  1.776604e-01     3.733879e-01           0                    0  2.6789352
#> 60  1.899537e-01     3.854038e-01           0                    0  2.5618503
#> 61  1.922471e-01     3.854038e-01           0                    0  4.8167273
#> 62  1.927019e-01     3.854038e-01           0                    0  2.5371943
#> 63  2.023614e-01     3.982987e-01           0                    0  2.4544775
#> 64  2.254763e-01     4.357730e-01           0                    0  4.0117172
#> 65  2.284294e-01     4.357730e-01           0                    0  3.9516418
#> 66  2.400637e-01     4.454554e-01           0                    0  2.1798848
#> 67  2.430289e-01     4.454554e-01           0                    0  3.6762963
#> 68  2.442820e-01     4.454554e-01           0                    0  2.1530802
#> 69  2.487921e-01     4.471046e-01           0                    0  3.5765766
#> 70  2.545120e-01     4.481030e-01           0                    0  3.4821053
#> 71  2.573558e-01     4.481030e-01           0                    0  3.4367100
#> 72  2.601889e-01     4.481030e-01           0                    0  3.3924786
#> 73  2.699582e-01     4.552048e-01           0                    0  1.6936383
#> 74  2.781015e-01     4.552048e-01           0                    0  1.9600080
#> 75  2.852114e-01     4.552048e-01           0                    0  3.0401533
#> 76  2.879396e-01     4.552048e-01           0                    0  3.0054545
#> 77  2.906576e-01     4.552048e-01           0                    0  2.9715356
#> 78  2.933652e-01     4.552048e-01           0                    0  2.9383704
#> 79  2.960627e-01     4.552048e-01           0                    0  2.9059341
#> 80  2.960627e-01     4.552048e-01           0                    0  2.9059341
#> 81  3.067514e-01     4.552048e-01           0                    0  2.7830175
#> 82  3.076467e-01     4.552048e-01           0                    0  1.8171048
#> 83  3.093985e-01     4.552048e-01           0                    0  2.7538889
#> 84  3.093985e-01     4.552048e-01           0                    0  2.7538889
#> 85  3.120355e-01     4.552048e-01           0                    0  2.7253608
#> 86  3.172799e-01     4.572461e-01           0                    0  2.6700337
#> 87  3.224848e-01     4.572461e-01           0                    0  2.6168977
#> 88  3.276505e-01     4.572461e-01           0                    0  2.5658252
#> 89  3.281847e-01     4.572461e-01           0                    0  1.5096513
#> 90  3.378657e-01     4.606299e-01           0                    0  2.4694081
#> 91  3.479277e-01     4.606299e-01           0                    0  2.3799399
#> 92  3.479277e-01     4.606299e-01           0                    0  2.3799399
#> 93  3.504195e-01     4.606299e-01           0                    0  2.3585714
#> 94  3.504195e-01     4.606299e-01           0                    0  2.3585714
#> 95  3.529019e-01     4.606299e-01           0                    0  2.3375811
#> 96  3.772170e-01     4.872386e-01           0                    0  2.1464499
#> 97  3.890341e-01     4.905620e-01           0                    0  2.0620833
#> 98  3.890341e-01     4.905620e-01           0                    0  2.0620833
#> 99  3.936988e-01     4.905620e-01           0                    0  2.0301538
#> 100 3.960179e-01     4.905620e-01           0                    0  2.0145547
#> 101 4.038144e-01     4.905620e-01           0                    0  1.4605047
#> 102 4.074829e-01     4.905620e-01           0                    0  1.9400000
#> 103 4.074829e-01     4.905620e-01           0                    0  1.9400000
#> 104 4.209577e-01     5.005837e-01           0                    0  1.8574648
#> 105 4.319553e-01     5.005837e-01           0                    0  1.7938322
#> 106 4.319553e-01     5.005837e-01           0                    0  1.7938322
#> 107 4.319553e-01     5.005837e-01           0                    0  1.7938322
#> 108 4.362964e-01     5.009329e-01           0                    0  1.7695749
#> 109 4.512340e-01     5.133304e-01           0                    0  1.6895726
#> 110 4.718981e-01     5.316782e-01           0                    0  1.5869880
#> 111 4.759378e-01     5.316782e-01           0                    0  1.5679365
#> 112 5.090666e-01     5.586218e-01           0                    0  1.4226306
#> 113 5.090666e-01     5.586218e-01           0                    0  1.4226306
#> 114 5.293925e-01     5.758305e-01           0                    0  1.3420408
#> 115 5.401270e-01     5.773771e-01           0                    0  1.3017822
#> 116 5.401270e-01     5.773771e-01           0                    0  1.3017822
#> 117 5.659174e-01     5.997757e-01           0                    0  1.2108756
#> 118 5.886982e-01     6.134334e-01           0                    0  1.1366811
#> 119 5.886982e-01     6.134334e-01           0                    0  1.1366811
#> 120 6.011817e-01     6.212211e-01           0                    0  1.0981869
#> 121 6.147791e-01     6.293601e-01           0                    0  1.0578495
#> 122 6.192091e-01     6.293601e-01           0                    0  1.0450465
#> 123 6.908733e-01     6.964901e-01           0                    0  0.8576612
#> 124 7.193720e-01     7.193720e-01           0                    0  0.7916768
#>     Combined.Score                                          Genes
#> 1      496.6971266 FCER2;HLA-DMA;IL1B;CD19;ITGA2B;CD14;CD1C;MS4A1
#> 2      157.4942918                              HLA-DMA;IL1B;GZMB
#> 3      152.1274703                              HLA-DMA;IL1B;GZMB
#> 4       58.7528355                   FCGR3A;HLA-DMA;IL1B;CD14;BID
#> 5       63.4917729                            FCGR3A;HLA-DMA;IL1B
#> 6       58.7278355                              CD79B;CD19;LILRA3
#> 7       46.5741304                             IL1B;S100A8;IL17RA
#> 8       45.8052573                            FCGR3A;HLA-DMA;FPR1
#> 9       30.4883714                       FCGR3A;ITGA2B;VDAC3;FPR1
#> 10      94.3396393                                 HLA-DMA;FCER1A
#> 11      40.9337495                                 IL1B;CD14;CD1C
#> 12      39.6881925                                 IL1B;CCL4;CD14
#> 13      39.6881925                                 IL1B;CCL4;CD14
#> 14      27.1877438                         FCER2;HLA-DMA;CD19;BID
#> 15      70.0536893                                   HLA-DMA;GZMB
#> 16      31.9514685                               FCER1A;BID;S1PR4
#> 17      22.5007466                            IL1B;CCL4;BID;GNG11
#> 18      28.7278988                             FCGR3A;IL1B;LILRA3
#> 19      27.2969728                                FCGR3A;GZMB;BID
#> 20      42.7098384                                   HLA-DMA;GZMB
#> 21      38.2567922                                      IL1B;CD14
#> 22      21.2819495                            FCGR3A;HLA-DMA;CD14
#> 23      35.3821120                                    HLA-DMA;BID
#> 24      19.7094217                                 IL1B;VDAC3;BID
#> 25      32.8351999                                      IL1B;CCL4
#> 26      31.2941854                                   HLA-DMA;IL1B
#> 27      13.6814864                          IL1B;CCL4;CCR7;IL17RA
#> 28      17.2087329                               HLA-DMA;IL1B;BID
#> 29      24.5374866                                      IL1B;CD14
#> 30      14.1768919                                CCL4;CCR7;GNG11
#> 31       9.5771778                        TCL1A;CD19;ITGA2B;GNG11
#> 32      11.5539998                                  IL1B;CD14;BID
#> 33      17.7856301                                   HLA-DMA;IL1B
#> 34      15.8036792                                      CCL4;CCR7
#> 35      14.1369007                                   HLA-DMA;IL1B
#> 36       9.5361862                                 FCGR3A;HLA-DMA
#> 37       9.0641742                                    IL1B;ITGA2B
#> 38       9.0641742                                       IL1B;BID
#> 39       8.7320888                                       GZMB;BID
#> 40      20.3572579                                         TALDO1
#> 41       7.4783993                                       IL1B;BID
#> 42       7.3928857                                  VDAC3;ZFP36L1
#> 43       4.6106392                         HLA-DMA;IL1B;BID;SRSF7
#> 44      16.2286503                                           NT5C
#> 45       6.9079518                                      VDAC3;BID
#> 46      14.9432755                                           IL1B
#> 47      14.3607166                                           CD19
#> 48       4.6786526                               FPR1;S1PR4;PTGDR
#> 49      12.8144994                                          VDAC3
#> 50       5.6386575                                     IL1B;VDAC3
#> 51      11.9255625                                          BLVRA
#> 52      11.1305110                                           TAF7
#> 53       3.9517238                                 IL1B;VDAC3;BID
#> 54       5.0493231                                      GZMB;CD14
#> 55       5.0000618                                      BID;GNG11
#> 56      10.0854555                                        HLA-DMA
#> 57       9.4717853                                          VDAC3
#> 58       9.4717853                                           IL1B
#> 59       4.6288830                                   ITGA2B;PARVB
#> 60       4.2551685                                    ITGA2B;FPR1
#> 61       7.9426568                                           NT5C
#> 62       4.1777717                                      BID;GNG11
#> 63       3.9215185                                  HLA-DMA;VDAC3
#> 64       5.9756141                                           CD14
#> 65       5.8347105                                         FCER1A
#> 66       3.1103704                                      IL1B;CD14
#> 67       5.2003963                                            BID
#> 68       3.0346201                                      IL1B;CD14
#> 69       4.9755105                                          ASGR1
#> 70       4.7649386                                         ITGA2B
#> 71       4.6646321                                        HLA-DMA
#> 72       4.5674552                                       MPHOSPH6
#> 73       2.2177995                                 IL1B;VDAC3;BID
#> 74       2.5083574                                     IL1B;VDAC3
#> 75       3.8139466                                         ITGA2B
#> 76       3.7418041                                          GNG11
#> 77       3.6716575                                         ITGA2B
#> 78       3.6034320                                          GNG11
#> 79       3.5370563                                         ITGA2B
#> 80       3.5370563                                        HLA-DMA
#> 81       3.2887409                                         ITGA2B
#> 82       2.1420090                                      IL1B;CD14
#> 83       3.2306569                                          GNG11
#> 84       3.2306569                                         FCGR3A
#> 85       3.1740594                                           IL1B
#> 86       3.0651211                                           IL1B
#> 87       2.9615412                                           IL1B
#> 88       2.8629674                                           IL1B
#> 89       1.6820213                               ITGA2B;BID;GNG11
#> 90       2.6795714                                           POP7
#> 91       2.5126468                                           IL1B
#> 92       2.5126468                                        HLA-DMA
#> 93       2.4732551                                          GNG11
#> 94       2.4732551                                          GNG11
#> 95       2.4347428                                          GNG11
#> 96       2.0926484                                         ITGA2B
#> 97       1.9467885                                           NT5C
#> 98       1.9467885                                          GNG11
#> 99       1.8924468                                          S1PR4
#> 100      1.8660736                                          GNG11
#> 101      1.3243856                                      BID;SRSF7
#> 102      1.7416472                                          GNG11
#> 103      1.7416472                                           IL1B
#> 104      1.6071212                                          VDAC3
#> 105      1.5058023                                        HLA-DMA
#> 106      1.5058023                                         FCER1A
#> 107      1.5058023                                          GNG11
#> 108      1.4677447                                          SRSF7
#> 109      1.3445100                                            BID
#> 110      1.1918157                                          VDAC3
#> 111      1.1641429                                           CD1C
#> 112      0.9605267                                          GNG11
#> 113      0.9605267                                           POP7
#> 114      0.8535717                                           IL1B
#> 115      0.8018341                                          VDAC3
#> 116      0.8018341                                          VDAC3
#> 117      0.6893601                                         ITGA2B
#> 118      0.6022610                                           IL1B
#> 119      0.6022610                                          GNG11
#> 120      0.5588213                                          VDAC3
#> 121      0.5146355                                          VDAC3
#> 122      0.5009036                                          VPS28
#> 123      0.3171621                                          VDAC3
#> 124      0.2607599                                         ITGA2B
#>                  db
#> 1   KEGG_2021_Human
#> 2   KEGG_2021_Human
#> 3   KEGG_2021_Human
#> 4   KEGG_2021_Human
#> 5   KEGG_2021_Human
#> 6   KEGG_2021_Human
#> 7   KEGG_2021_Human
#> 8   KEGG_2021_Human
#> 9   KEGG_2021_Human
#> 10  KEGG_2021_Human
#> 11  KEGG_2021_Human
#> 12  KEGG_2021_Human
#> 13  KEGG_2021_Human
#> 14  KEGG_2021_Human
#> 15  KEGG_2021_Human
#> 16  KEGG_2021_Human
#> 17  KEGG_2021_Human
#> 18  KEGG_2021_Human
#> 19  KEGG_2021_Human
#> 20  KEGG_2021_Human
#> 21  KEGG_2021_Human
#> 22  KEGG_2021_Human
#> 23  KEGG_2021_Human
#> 24  KEGG_2021_Human
#> 25  KEGG_2021_Human
#> 26  KEGG_2021_Human
#> 27  KEGG_2021_Human
#> 28  KEGG_2021_Human
#> 29  KEGG_2021_Human
#> 30  KEGG_2021_Human
#> 31  KEGG_2021_Human
#> 32  KEGG_2021_Human
#> 33  KEGG_2021_Human
#> 34  KEGG_2021_Human
#> 35  KEGG_2021_Human
#> 36  KEGG_2021_Human
#> 37  KEGG_2021_Human
#> 38  KEGG_2021_Human
#> 39  KEGG_2021_Human
#> 40  KEGG_2021_Human
#> 41  KEGG_2021_Human
#> 42  KEGG_2021_Human
#> 43  KEGG_2021_Human
#> 44  KEGG_2021_Human
#> 45  KEGG_2021_Human
#> 46  KEGG_2021_Human
#> 47  KEGG_2021_Human
#> 48  KEGG_2021_Human
#> 49  KEGG_2021_Human
#> 50  KEGG_2021_Human
#> 51  KEGG_2021_Human
#> 52  KEGG_2021_Human
#> 53  KEGG_2021_Human
#> 54  KEGG_2021_Human
#> 55  KEGG_2021_Human
#> 56  KEGG_2021_Human
#> 57  KEGG_2021_Human
#> 58  KEGG_2021_Human
#> 59  KEGG_2021_Human
#> 60  KEGG_2021_Human
#> 61  KEGG_2021_Human
#> 62  KEGG_2021_Human
#> 63  KEGG_2021_Human
#> 64  KEGG_2021_Human
#> 65  KEGG_2021_Human
#> 66  KEGG_2021_Human
#> 67  KEGG_2021_Human
#> 68  KEGG_2021_Human
#> 69  KEGG_2021_Human
#> 70  KEGG_2021_Human
#> 71  KEGG_2021_Human
#> 72  KEGG_2021_Human
#> 73  KEGG_2021_Human
#> 74  KEGG_2021_Human
#> 75  KEGG_2021_Human
#> 76  KEGG_2021_Human
#> 77  KEGG_2021_Human
#> 78  KEGG_2021_Human
#> 79  KEGG_2021_Human
#> 80  KEGG_2021_Human
#> 81  KEGG_2021_Human
#> 82  KEGG_2021_Human
#> 83  KEGG_2021_Human
#> 84  KEGG_2021_Human
#> 85  KEGG_2021_Human
#> 86  KEGG_2021_Human
#> 87  KEGG_2021_Human
#> 88  KEGG_2021_Human
#> 89  KEGG_2021_Human
#> 90  KEGG_2021_Human
#> 91  KEGG_2021_Human
#> 92  KEGG_2021_Human
#> 93  KEGG_2021_Human
#> 94  KEGG_2021_Human
#> 95  KEGG_2021_Human
#> 96  KEGG_2021_Human
#> 97  KEGG_2021_Human
#> 98  KEGG_2021_Human
#> 99  KEGG_2021_Human
#> 100 KEGG_2021_Human
#> 101 KEGG_2021_Human
#> 102 KEGG_2021_Human
#> 103 KEGG_2021_Human
#> 104 KEGG_2021_Human
#> 105 KEGG_2021_Human
#> 106 KEGG_2021_Human
#> 107 KEGG_2021_Human
#> 108 KEGG_2021_Human
#> 109 KEGG_2021_Human
#> 110 KEGG_2021_Human
#> 111 KEGG_2021_Human
#> 112 KEGG_2021_Human
#> 113 KEGG_2021_Human
#> 114 KEGG_2021_Human
#> 115 KEGG_2021_Human
#> 116 KEGG_2021_Human
#> 117 KEGG_2021_Human
#> 118 KEGG_2021_Human
#> 119 KEGG_2021_Human
#> 120 KEGG_2021_Human
#> 121 KEGG_2021_Human
#> 122 KEGG_2021_Human
#> 123 KEGG_2021_Human
#> 124 KEGG_2021_Human
# }